# backend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QuickFlow API",
    "description": "Single-user productivity API: tasks, habits, learning resources, plans, dashboard. Errors are RFC 7807 problem+json; validation errors list {field, message} in `errors`.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/plans": {
      "get": {
        "tags": [
          "Plans"
        ],
        "summary": "List plans",
        "description": "ACTIVE: not started or in progress, by priority. COMPLETED: history, latest first.",
        "operationId": "list_1",
        "parameters": [
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "ALL",
              "enum": [
                "ACTIVE",
                "COMPLETED",
                "ALL"
              ]
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Plan or item not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Plan"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Plan"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Plans"
        ],
        "summary": "Create a plan from existing tasks, habits and learning resources",
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Plan or item not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/{id}/start-notification/ack": {
      "post": {
        "tags": [
          "Plans"
        ],
        "summary": "Acknowledge a plan's start notification",
        "operationId": "acknowledgeStart",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Plan or item not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          },
          "409": {
            "description": "Plan has not started yet",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/plans/{id}/items/{itemId}": {
      "patch": {
        "tags": [
          "Plans"
        ],
        "summary": "Mark a plan item done or not done",
        "description": "A task item also completes/reopens the task; a habit item records/removes today's completion; a learning item changes only the plan.",
        "operationId": "setItemDone",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanItemUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Plan or item not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          },
          "409": {
            "description": "The habit behind the item is inactive",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/plans/{id}": {
      "get": {
        "tags": [
          "Plans"
        ],
        "summary": "Get a plan with progress and rest time",
        "operationId": "get_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Plan or item not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Plan"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Plans"
        ],
        "summary": "Remove a plan",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Plan or item not found"
          },
          "204": {
            "description": "Deleted"
          },
          "400": {
            "description": "Validation failed"
          }
        }
      }
    },
    "/api/plans/start-notifications": {
      "get": {
        "tags": [
          "Plans"
        ],
        "summary": "Plans that have started and not been acknowledged yet",
        "operationId": "startNotifications",
        "responses": {
          "404": {
            "description": "Plan or item not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Plan"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Plan"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/plans/sources": {
      "get": {
        "tags": [
          "Plans"
        ],
        "summary": "Tasks, habits and learning resources that can be added to a plan",
        "operationId": "sources",
        "responses": {
          "404": {
            "description": "Plan or item not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlanSources"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlanSources"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PlanItemRef": {
        "required": [
          "sourceId",
          "sourceType"
        ],
        "type": "object",
        "properties": {
          "sourceType": {
            "type": "string",
            "enum": [
              "TASK",
              "HABIT",
              "LEARNING_RESOURCE"
            ]
          },
          "sourceId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "description": "An existing task, habit or learning card"
      },
      "PlanRequest": {
        "required": [
          "endDateTime",
          "estimatedMinutes",
          "items",
          "priorityOrder",
          "startDateTime",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanItemRef"
            }
          },
          "estimatedMinutes": {
            "maximum": 100000,
            "minimum": 1,
            "type": "integer",
            "description": "Estimated duration in minutes",
            "format": "int32"
          },
          "startDateTime": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-24T09:00:00+03:00"
          },
          "endDateTime": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-24T11:00:00+03:00"
          },
          "priorityOrder": {
            "minimum": 1,
            "type": "integer",
            "description": "1 = highest priority",
            "format": "int32"
          }
        }
      },
      "Plan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "estimatedMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "startDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "endDateTime": {
            "type": "string",
            "format": "date-time"
          },
          "priorityOrder": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanItem"
            }
          },
          "doneCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "progressPercent": {
            "type": "integer",
            "format": "int32"
          },
          "restSeconds": {
            "type": "integer",
            "description": "Seconds until end while IN_PROGRESS, otherwise null",
            "format": "int64"
          },
          "startNotified": {
            "type": "boolean"
          },
          "serverTime": {
            "type": "string",
            "description": "Server time used for the computation; clients count down from it",
            "format": "date-time"
          }
        }
      },
      "PlanItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "sourceType": {
            "type": "string",
            "enum": [
              "TASK",
              "HABIT",
              "LEARNING_RESOURCE"
            ]
          },
          "sourceId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "done": {
            "type": "boolean"
          },
          "sourceAvailable": {
            "type": "boolean",
            "description": "false when the source was deleted (I-5)"
          }
        }
      },
...
      "PlanItemUpdate": {
        "required": [
          "done"
        ],
        "type": "object",
        "properties": {
          "done": {
            "type": "boolean"
          }
        }
      },
...
      "PlanSource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "PlanSources": {
        "type": "object",
        "properties": {
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanSource"
            }
          },
          "habits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanSource"
            }
          },
          "learningResources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanSource"
            }
          }
        },
        "description": "Items that can be added to a plan (I-18)"
      },
...
    }
  }
}